R CMD check says that install_conversion_constant is deprecated and to use install_unit instead.
Like the last checkin, David will have to review this one to make sure it is correct. My reading of the documentation is that "veh mi" is translated as multiplication. And since the offset in the old code was 1, I think that this should work. But I really don't know how to test this code.